f1-producer/requirements.txt
kafka-python
fastf1
pandas